// packages/core/src/core/enums.ts
export enum StatusCondition {
    NONE = "none",
    POISON = "poison",
    SLEEP = "sleep",
    PARALYSIS = "paralysis",
    BURN = "burn",
    FREEZE = "freeze",
    CONFUSION = "confusion",
}

export enum AILayer {
    AI_BASIC = "AI_BASIC",
    AI_SETUP = "AI_SETUP",
    AI_TYPES = "AI_TYPES",
    AI_OFFENSIVE = "AI_OFFENSIVE",
    AI_SMART = "AI_SMART",
    AI_OPPORTUNIST = "AI_OPPORTUNIST",
    AI_AGGRESSIVE = "AI_AGGRESSIVE",
    AI_CAUTIOUS = "AI_CAUTIOUS",
    AI_STATUS = "AI_STATUS",
    AI_RISKY = "AI_RISKY",
    AI_NONE = "AI_NONE",
}

export enum BattleScene {
    ON = "on",
    OFF = "off",
}

export enum BattleStyle {
    SET = "set",
    SHIFT = "shift",
}

export enum ItemPocket {
    ITEM = "ITEM",
    BALL = "BALL",
    KEY_ITEM = "KEY_ITEM",
    TM_HM = "TM_HM",
}

export enum ItemEffect {
    NONE = "NONE",
    POKE_BALL = "POKE_BALL",
    TOWN_MAP = "TOWN_MAP",
    BICYCLE = "BICYCLE",
    EVO_STONE = "EVO_STONE",
    STATUS_HEAL = "STATUS_HEAL",
    FULL_RESTORE = "FULL_RESTORE",
    RESTORE_HP = "RESTORE_HP",
    ESCAPE_ROPE = "ESCAPE_ROPE",
    REPEL = "REPEL",
    RESTORE_PP = "RESTORE_PP",
    VITAMIN = "VITAMIN",
    RARE_CANDY = "RARE_CANDY",
    X_ACCURACY = "X_ACCURACY",
    POKE_DOLL = "POKE_DOLL",
    REVIVE = "REVIVE",
    GUARD_SPEC = "GUARD_SPEC",
    SUPER_REPEL = "SUPER_REPEL",
    MAX_REPEL = "MAX_REPEL",
    DIRE_HIT = "DIRE_HIT",
    X_ITEM = "X_ITEM",
    COIN_CASE = "COIN_CASE",
    ITEMFINDER = "ITEMFINDER",
    POKE_FLUTE = "POKE_FLUTE",
    OLD_ROD = "OLD_ROD",
    GOOD_ROD = "GOOD_ROD",
    SUPER_ROD = "SUPER_ROD",
    BITTER_BERRY = "BITTER_BERRY",
    ENERGY_POWDER = "ENERGY_POWDER",
    ENERGY_ROOT = "ENERGY_ROOT",
    HEAL_POWDER = "HEAL_POWDER",
    REVIVAL_HERB = "REVIVAL_HERB",
    CARD_KEY = "CARD_KEY",
    BASEMENT_KEY = "BASEMENT_KEY",
    SACRED_ASH = "SACRED_ASH",
    NORMAL_BOX = "NORMAL_BOX",
    GORGEOUS_BOX = "GORGEOUS_BOX",
    SQUIRTBOTTLE = "SQUIRTBOTTLE",
    BLUE_CARD = "BLUE_CARD",
    TRADE = "TRADE",
    HAPPINESS = "HAPPINESS",
    STAT = "STAT",
}